# Reuse bits in RejectionSampler