I thought I tried this in the past, but actually probably I messed up something, I see a good speed-up in many cases thanks to this.