@Tortar 2025-01-09
e.g.

julia> using DynamicDiscreteSamplers, BenchmarkTools, Random

julia> function setup(rng, indices)
           ds = DynamicDiscreteSampler()
           append!(ds, indices, rand(rng, length(indices)))
           return ds
       end
setup (generic function with 1 method)

julia> rng = Xoshiro(42)
Xoshiro(0xa379de7eeeb2a4e8, 0x953dccb6b532b3af, 0xf597b8ff8cfd652a, 0xccd7337c571680d1, 0xc90c4a0730db3f7e)

julia> ds = setup(rng, 1:10^7);

julia> @benchmark rand($rng, $ds, 10^4)

julia> @benchmark rand($rng, $ds, 10^4)
# main
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  396.247 μs …  17.252 ms  ┊ GC (min … max): 0.00% … 97.34%
 Time  (median):     414.120 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   425.965 μs ± 173.868 μs  ┊ GC (mean ± σ):  0.59% ±  2.08%

         ▂▅▇███▅▃▁                      ▁▁▁                      
  ▁▁▁▂▃▅▇█████████▇▅▄▃▂▂▂▂▂▂▂▁▁▁▁▂▂▃▄▅▇▇████▇▅▄▃▃▂▂▂▂▁▂▁▁▁▁▁▁▁▁ ▃
  396 μs           Histogram: frequency by time          470 μs <

 Memory estimate: 78.69 KiB, allocs estimate: 7.
#pr
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  384.749 μs …  17.567 ms  ┊ GC (min … max): 0.00% … 97.45%
 Time  (median):     403.103 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   415.721 μs ± 177.786 μs  ┊ GC (mean ± σ):  0.62% ±  2.08%

         ▂▄██▇▅▃▁                                                
  ▁▁▁▂▂▅▇████████▆▅▄▃▂▂▂▂▂▂▁▁▁▁▁▂▃▄▆▇███▇▇▆▄▃▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁ ▃
  385 μs           Histogram: frequency by time          466 μs <

 Memory estimate: 78.69 KiB, allocs estimate: 7.

julia> ds = setup(rng, 1:10^3);

julia> @benchmark rand($rng, $ds, 10^4)

#main
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  117.641 μs …  2.286 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     124.695 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   127.317 μs ± 31.320 μs  ┊ GC (mean ± σ):  1.23% ± 5.00%

          ▃▇█▇▃                                                 
  ▂▁▂▂▂▃▅▇█████▇▆▆▆▆▅▄▄▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▁▁▁▂▁▂▁▁▂▁▁▁▂ ▃
  118 μs          Histogram: frequency by time          155 μs <

 Memory estimate: 78.50 KiB, allocs estimate: 7.

#pr
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  103.687 μs … 809.106 μs  ┊ GC (min … max): 0.00% … 83.17%
 Time  (median):     108.737 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   110.606 μs ±  21.333 μs  ┊ GC (mean ± σ):  1.30% ±  5.04%

                 ▁▃▃▆█▇▇▇▆▅▃▂                                    
  ▁▁▁▁▁▂▂▂▂▃▄▄▅▇▇█████████████▇▆▅▅▄▃▃▃▃▃▄▃▃▃▂▃▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  104 μs           Histogram: frequency by time          118 μs <

 Memory estimate: 78.50 KiB, allocs estimate: 7.