@Tortar 2025-01-09
actually what we can do is reusing those bits only in bulk generation, which wouldn't make any overhead